Fix to back button issue -- "Back button issue: If you use the up-navigation, the activity seems to stay in memory. Press the back button and you go through the activities that are still in memory due to using up-navigation."
I noticed that this was being caused due to starting a new instance of
the TimelineScreen activity, by doing such, the prior activity stays in
memory. This method also didn't allow up navigation to function as it
should, as it would just jump back to the TimelineScreen if you are deep
in the Settings menus, for example.