Welp, here we go soooo..
- Profile Info Tab is now 10000% better
- British language 
- Messages, About, Search T/U, Trends fixed.

Search has been stripped of duplicated fragments. We do not need 2 that do the same job (search twitter for Tweets)

Also, the adapter for profiles is gone because it was messy and we were create a virtual list. ListAdapters are meant for a list of data, we didn't really have one that was well, proper. This is more optimized and simple :)